+++ PSiVR/PSiVR/Geometry/quat.cpp
@@ -1,19 +1,19 @@
 
 #include "quat.h"
 
-const quat operator+(const float& t, const quat& q)
+const quat operator+(const long double& t, const quat& q)
 {
 	return q + t;
 }
-const quat operator-(const float& t, const quat& q)
+const quat operator-(const long double& t, const quat& q)
 {
 	return q - t;
 }
-const quat operator*(const float& t, const quat& q)
+const quat operator*(const long double& t, const quat& q)
 {
 	return q * t;
 }
-const quat operator/(const float& t, const quat& q)
+const quat operator/(const long double& t, const quat& q)
 {
 	return q / t;
 }
